+++ components/spider-storage/src/state.rs
@@ -7,4 +7,4 @@ pub use job_cache::JobCache;
 pub use service::ServiceState;
 
 #[cfg(test)]
-mod test_mocks;
+mod test_utils;
